requirements.txt
tensorflow-deps==2.10.0         # Apple M-series dependencies
tensorflow-macos==2.10.0        # Core TensorFlow for M1/M2
tensorflow-metal==0.7.0         # GPU acceleration
tensorflow-privacy==0.7.0       # Differential privacy
numpy==1.23.5                   # Fixed for M1 compatibility
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
plotly==5.15.0
psutil==5.9.5
seaborn>=0.11.0
